# improve trace_context performance