Propagation is essential logic, regardless of whether it hits sampling. In the past, the implementation of Propagation did not prioritize memory management, allocating a lot of temporary heap memory. This caused its efficiency to be low. In some scenarios, too much CPU was wasted.
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/c9adedeb-f53a-408e-bfc8-36f7401e8d87)
For example, even if it doesn't hit sampling, startServerSpan consumes 25% of the entire request CPU.

Typical server span Propagation:
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/71ac2f19-5410-4cd4-9a5a-931dad4bb00e)

Typical client span Propagation
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/04aef19c-6cff-4202-b7b9-0873794ef2bb)

It can be seen that the poor performance of Propagation is mainly due to regular expressions and Sprintf.

This CR can significantly improve Propagation by optimizing it in the following ways:

Use string.Builder to pre-allocate memory
Use hex.Decode to a stack-based [N]byte array instead of hex.DecodeString, which temporarily allocates a heap-based string.
Use strings.Cut instead of strings.Split
Avoid using regex
All unit tests pass. Benchmarks: 

New logic:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/propagation
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkInject/SampledSpanContext-16            3489816               344.4 ns/op            96 B/op          3 allocs/op
BenchmarkInject/WithoutSpanContext-16           19206630                60.12 ns/op           16 B/op          1 allocs/op
BenchmarkExtract/Sampled-16                      2094048               576.1 ns/op           160 B/op          4 allocs/op
BenchmarkExtract/BogusVersion-16                 8635071               137.9 ns/op            16 B/op          1 allocs/op
BenchmarkExtract/FutureAdditionalData-16         2073709               577.8 ns/op           160 B/op          4 allocs/op
PASS
ok      go.opentelemetry.io/otel/propagation    7.692s
```

Old logic:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/propagation
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkInject/SampledSpanContext-16            1289794               926.0 ns/op           224 B/op         11 allocs/op
BenchmarkInject/WithoutSpanContext-16           18479284                60.12 ns/op           16 B/op          1 allocs/op
BenchmarkExtract/Sampled-16                       760831              1423 ns/op             320 B/op          6 allocs/op
BenchmarkExtract/BogusVersion-16                 6141056               191.4 ns/op            16 B/op          1 allocs/op
BenchmarkExtract/FutureAdditionalData-16          704563              1586 ns/op             320 B/op          6 allocs/op
PASS
ok      go.opentelemetry.io/otel/propagation    6.938s
```

It can be seen that the performance has improved by about 2-3 times.