@xiehuc 2023-11-17
> Similar to #4722, please use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to make comparisons between benchmarks in your description.
> 
> Also, the increases in allocations is concerning. The impact on GC is not captured in the benchmarks but will be when it is run on real systems.

allocation is on stack, less GC than old implement,