@xiehuc 2023-11-17
original benchmark result
[old.txt](https://github.com/open-telemetry/opentelemetry-go/files/13386245/old.txt)
[new.txt](https://github.com/open-telemetry/opentelemetry-go/files/13386246/new.txt)