@MrAlias 2023-11-17
It seems like I was confusing old vs new in my last comment. The benchstat output was helpful in clarifying this.