@xiehuc 2023-11-29
newest code benchmark result
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/propagation
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
                                │ /tmp/old.txt │            /tmp/new2.txt            │
                                │    sec/op    │   sec/op     vs base                │
Inject/SampledSpanContext-16       916.0n ± 0%   335.7n ± 1%  -63.35% (p=0.000 n=10)
Inject/WithoutSpanContext-16       61.79n ± 1%   60.83n ± 0%   -1.55% (p=0.000 n=10)
Extract/Sampled-16                1398.0n ± 0%   586.4n ± 1%  -58.05% (p=0.000 n=10)
Extract/BogusVersion-16            190.4n ± 0%   137.0n ± 0%  -28.07% (p=0.000 n=10)
Extract/FutureAdditionalData-16   1553.0n ± 0%   579.0n ± 0%  -62.72% (p=0.000 n=10)
geomean                            471.9n        248.6n       -47.32%

                                │ /tmp/old.txt │            /tmp/new2.txt             │
                                │     B/op     │    B/op     vs base                  │
Inject/SampledSpanContext-16       224.00 ± 0%   96.00 ± 0%  -57.14% (p=0.000 n=10)
Inject/WithoutSpanContext-16        16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=10) ¹
Extract/Sampled-16                  320.0 ± 0%   160.0 ± 0%  -50.00% (p=0.000 n=10)
Extract/BogusVersion-16             16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=10) ¹
Extract/FutureAdditionalData-16     320.0 ± 0%   160.0 ± 0%  -50.00% (p=0.000 n=10)
geomean                             89.90        57.51       -36.03%
¹ all samples are equal

                                │ /tmp/old.txt │            /tmp/new2.txt             │
                                │  allocs/op   │ allocs/op   vs base                  │
Inject/SampledSpanContext-16       11.000 ± 0%   3.000 ± 0%  -72.73% (p=0.000 n=10)
Inject/WithoutSpanContext-16        1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
Extract/Sampled-16                  6.000 ± 0%   4.000 ± 0%  -33.33% (p=0.000 n=10)
Extract/BogusVersion-16             1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
Extract/FutureAdditionalData-16     6.000 ± 0%   4.000 ± 0%  -33.33% (p=0.000 n=10)
geomean                             3.308        2.169       -34.43%
¹ all samples are equal

```