@xiehuc 2023-11-29
[new2.txt](https://github.com/open-telemetry/opentelemetry-go/files/13495381/new2.txt)